requirements.txt
python-dotenv==1.0.1
scrapegraphai==1.11.3

